ByteConvertible and Block0 in datafile

+++ src/queries/SequentialNNQuery.java
@@ -53,7 +53,7 @@ public class SequentialNNQuery extends Query{
     }
 
     private void search() {
-        int numBlocks = DataMetaData.getNumberOfBlocks();
+        long numBlocks = DataMetaData.getNumberOfBlocks();
 
         for (int blockId = 0; blockId < numBlocks; blockId++) {
             ArrayList<Record> blockRecords = FileHandler.getDataBlock(blockId);
