+++ src/queries/SequentialRangeQuery.java
@@ -38,7 +38,7 @@ public class SequentialRangeQuery extends Query{
     }
 
     private void search() {
-        int numBlocks = DataMetaData.getNumberOfBlocks();
+        long numBlocks = DataMetaData.getNumberOfBlocks();
 
         for (int blockId = 0; blockId < numBlocks; blockId++) {
             ArrayList<Record> blockRecords = FileHandler.getDataBlock(blockId);
